wip: created a composable for the application page
